# docs/swagger.yaml
basePath: /
definitions:
  handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
  models.AddictionWithoutTime:
    properties:
      admin_name:
        type: string
      reason:
        type: string
      user_name:
        type: string
      user_phone:
        type: string
    type: object
  models.Name:
    properties:
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for BlackListApi Application
  title: BlackListApi App API
  version: "1.0"
paths:
  /add:
    post:
      consumes:
      - application/json
      description: Add user to blacklist
      operationId: add-into-blacklist
      parameters:
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddictionWithoutTime'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add user to blacklist
      tags:
      - Addition
  /auth:
    post:
      consumes:
      - application/json
      description: generate token
      operationId: create-account
      parameters:
      - description: token generating by name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/models.Name'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Auth
      tags:
      - auth
  /delete:
    get:
      consumes:
      - application/json
      description: Remove user from blacklist
      operationId: delete-from-blacklist
      parameters:
      - description: ID of user to remove from blacklist
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove user from blacklist
      tags:
      - Removal
  /search:
    get:
      consumes:
      - application/json
      description: Search users in blacklist
      operationId: search-users
      parameters:
      - description: user phone
        in: query
        name: number
        type: string
      - description: user name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search users in blacklist
      tags:
      - Search
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"